requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic>=2.11.7
pydantic-settings
requests
alembic
supabase
sqlmodel
httpx>=0.26
psycopg2-binary
pytesseract
pillow
google-genai